Share navbar build menu
