+++ docs/app/chat/chat-page.module.css
@@ -104,133 +104,10 @@
   outline-offset: 1px;
 }
 
-.ctaWrap {
+.buildForFreeMenu {
   position: relative;
   min-width: 0;
   justify-self: end;
-}
-
-.ctaButton {
-  display: flex;
-  box-sizing: border-box;
-  height: 2.25rem;
-  align-items: center;
-  gap: 0.375rem;
-  padding-inline: 0.625rem;
-  border: 0;
-  border-radius: 10px;
-  background: var(--home-surface-ink);
-  box-shadow: var(--home-bevel-on-dark);
-  color: #fff;
-  font-family: "Inter", "Segoe UI", sans-serif;
-  font-size: 0.9375rem;
-  font-weight: 500;
-  line-height: 1;
-  white-space: nowrap;
-  cursor: pointer;
-  transition: opacity 120ms ease;
-}
-
-.ctaButton:hover {
-  opacity: 0.88;
-}
-
-.ctaButton:focus-visible,
-.ctaMenuItem:focus-visible {
-  outline: 2px solid var(--openui-text-neutral-primary);
-  outline-offset: 2px;
-}
-
-.ctaArrow {
-  transition: transform 160ms ease;
-}
-
-.ctaArrow[data-open="true"] {
-  transform: rotate(90deg);
-}
-
-.ctaMenu {
-  position: absolute;
-  top: calc(100% + 8px);
-  right: 0;
-  z-index: 60;
-  width: 22.5rem;
-  opacity: 0;
-  visibility: hidden;
-  pointer-events: none;
-  transition:
-    opacity 0.18s ease,
-    visibility 0s linear 0.18s;
-}
-
-.ctaMenuOpen {
-  opacity: 1;
-  visibility: visible;
-  pointer-events: auto;
-  transition:
-    opacity 0.18s ease,
-    visibility 0s;
-}
-
-.ctaMenuCard {
-  display: flex;
-  flex-direction: column;
-  gap: 2px;
-  border: 1px solid var(--openui-border-default);
-  border-radius: 12px;
-  background: var(--openui-foreground);
-  padding: 0.375rem;
-  box-shadow: var(--openui-shadow-l);
-}
-
-.ctaMenuItem {
-  display: flex;
-  width: 100%;
-  min-height: 40px;
-  align-items: center;
-  justify-content: space-between;
-  gap: 0.5rem;
-  padding: 8px 12px;
-  border: 0;
-  border-radius: 8px;
-  background: transparent;
-  color: var(--openui-text-neutral-primary);
-  font-size: 0.8125rem;
-  line-height: 1.25rem;
-  white-space: nowrap;
-  cursor: pointer;
-  transition: background-color 120ms ease;
-}
-
-.ctaMenuItem:hover {
-  background: var(--openui-highlight) !important;
-}
-
-.ctaMenuItemLabel {
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.ctaMenuItemRunner {
-  color: var(--openui-text-neutral-primary);
-  font-weight: 600;
-}
-
-.ctaMenuItemIcon {
-  display: inline-flex;
-  width: 1.5rem;
-  height: 1.5rem;
-  flex-shrink: 0;
-  align-items: center;
-  justify-content: center;
-  color: var(--openui-text-neutral-secondary);
-  opacity: 0;
-  transition: opacity 0.12s ease;
-}
-
-.ctaMenuItem:hover .ctaMenuItemIcon,
-.ctaMenuItem:focus-visible .ctaMenuItemIcon {
-  opacity: 1;
 }
 
 .agentViewport {
@@ -453,7 +330,7 @@
     font-size: 0.6875rem !important;
   }
 
-  .ctaWrap {
+  .buildForFreeMenu {
     display: none;
   }
 }
@@ -465,8 +342,6 @@
 }
 
 @media (prefers-reduced-motion: reduce) {
-  .ctaArrow,
-  .ctaMenu,
   .modeItem {
     transition: none;
   }
