+++ docs/app/compare/_components/chat-page-header.tsx
@@ -1,7 +1,8 @@
 "use client";
 
-import { ClipboardCommandButton } from "@/app/(home)/components/Button/Button";
 import { GitHubButton } from "@/app/(home)/components/GitHubButton/GitHubButton";
+import { BuildForFreeMenu } from "@/app/_components/build-for-free-menu";
+import { useHeaderDropdown } from "@/app/_components/use-header-dropdown";
 import trayStyles from "@/components/site-marketing-header.module.css";
 import {
   ArrowLeft,
@@ -17,19 +18,11 @@ import {
 } from "lucide-react";
 import { useTheme } from "next-themes";
 import Link from "next/link";
-import { useEffect, useRef, useState, type ReactNode } from "react";
+import type { ReactNode } from "react";
 import styles from "../chat-page.module.css";
 import { getComparisonPair, type ComparisonPair } from "./chat-types";
 import type { ComparisonMode } from "./comparison-mode-controller";
 import { useHasMounted } from "./use-has-mounted";
-
-// Same runner set and order as the homepage hero's command dropdown.
-const CLI_COMMANDS = [
-  { id: "pnpm", runner: "pnpx", command: "pnpx @openuidev/cli@latest create" },
-  { id: "bun", runner: "bunx", command: "bunx @openuidev/cli@latest create" },
-  { id: "yarn", runner: "yarn dlx", command: "yarn dlx @openuidev/cli@latest create" },
-  { id: "npm", runner: "npx", command: "npx @openuidev/cli@latest create" },
-] as const;
 
 // Display titles for the pair options, mirroring the panel headings: mode name
 // plus a chip for the OpenUI flavor. aria-labels keep the canonical names.
@@ -138,62 +131,6 @@ function PairTitle({ pair }: { pair: ComparisonPair }) {
   );
 }
 
-// Shared open/close behavior for the header dropdowns: hover opens (mouse only,
-// with a grace delay so the pointer can cross into the menu), click toggles,
-// outside pointer-down and Escape close.
-function useHeaderDropdown() {
-  const [open, setOpen] = useState(false);
-  const wrapRef = useRef<HTMLDivElement | null>(null);
-  const triggerRef = useRef<HTMLButtonElement | null>(null);
-  const closeTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-  useEffect(() => {
-    return () => {
-      if (closeTimeoutRef.current) clearTimeout(closeTimeoutRef.current);
-    };
-  }, []);
-
-  const cancelScheduledClose = () => {
-    if (closeTimeoutRef.current) clearTimeout(closeTimeoutRef.current);
-    closeTimeoutRef.current = null;
-  };
-
-  const handleHoverOpen = (event: React.PointerEvent) => {
-    if (event.pointerType !== "mouse") return;
-    cancelScheduledClose();
-    setOpen(true);
-  };
-
-  const handleHoverClose = (event: React.PointerEvent) => {
-    if (event.pointerType !== "mouse") return;
-    cancelScheduledClose();
-    closeTimeoutRef.current = setTimeout(() => setOpen(false), 160);
-  };
-
-  useEffect(() => {
-    if (!open) return;
-
-    const handlePointerDown = (event: PointerEvent) => {
-      if (!wrapRef.current?.contains(event.target as Node)) setOpen(false);
-    };
-    const handleKeyDown = (event: KeyboardEvent) => {
-      if (event.key === "Escape") {
-        setOpen(false);
-        triggerRef.current?.focus();
-      }
-    };
-
-    document.addEventListener("pointerdown", handlePointerDown);
-    document.addEventListener("keydown", handleKeyDown);
-    return () => {
-      document.removeEventListener("pointerdown", handlePointerDown);
-      document.removeEventListener("keydown", handleKeyDown);
-    };
-  }, [open]);
-
-  return { open, setOpen, wrapRef, triggerRef, handleHoverOpen, handleHoverClose };
-}
-
 // Menu display order, independent of the default comparison pair.
 const PAIR_MENU_ORDER: readonly ComparisonPair[] = ["markdown-oss", "markdown-cloud", "oss-cloud"];
 const PAIR_MENU_OPTIONS = PAIR_MENU_ORDER.map((pair) => getComparisonPair(pair));
@@ -257,61 +194,6 @@ function PairSwitcher({
                 </span>
               </span>
             </button>
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function BuildForFreeMenu() {
-  const { open, setOpen, wrapRef, triggerRef, handleHoverOpen, handleHoverClose } =
-    useHeaderDropdown();
-
-  return (
-    <div
-      className={styles.ctaWrap}
-      ref={wrapRef}
-      onPointerEnter={handleHoverOpen}
-      onPointerLeave={handleHoverClose}
-    >
-      <button
-        type="button"
-        ref={triggerRef}
-        className={styles.ctaButton}
-        aria-haspopup="menu"
-        aria-expanded={open}
-        onClick={() => setOpen((value) => !value)}
-      >
-        <span>Build for free</span>
-        <ArrowRight
-          size={15}
-          strokeWidth={2}
-          aria-hidden="true"
-          className={styles.ctaChevron}
-          data-open={open}
-        />
-      </button>
-
-      <div className={`${styles.ctaMenu} ${open ? styles.ctaMenuOpen : ""}`.trim()}>
-        <div
-          className={styles.ctaMenuCard}
-          role="menu"
-          aria-label="Copy the setup command for a package manager"
-        >
-          {CLI_COMMANDS.map((item) => (
-            <ClipboardCommandButton
-              key={item.id}
-              command={item.command}
-              className={styles.ctaMenuItem}
-              iconContainerClassName={styles.ctaMenuItemIcon}
-              copyIconColor="currentColor"
-            >
-              <span className={styles.ctaMenuItemLabel}>
-                <span className={styles.ctaMenuItemRunner}>{item.runner}</span>
-                {item.command.slice(item.runner.length)}
-              </span>
-            </ClipboardCommandButton>
           ))}
         </div>
       </div>
@@ -541,7 +423,7 @@ export function ChatPageHeader({
 
         {rightHeading}
 
-        <BuildForFreeMenu />
+        <BuildForFreeMenu className={styles.buildForFreeMenu} />
 
         <MobileMenu pair={pair} onPairChange={onPairChange} onReset={onReset} />
       </div>
